// frontend/src/app/models/film.ts
export class Film{
    idF: number;
    naziv: string;
    datumPocetka: string;
    slika: string;
    zanr: String[];
    trajanje: string;
    trejler: string;
    sadrzaj: string;
    glumci: string;
    reziser: string;
    godIzlaska: string;
    brOcena: number;
    ocena: number;
    zbir: number;
}